removed posenet class to reduce complexity
